+++ src/app/index.js
@@ -1,13 +1,13 @@
 import { createApp } from './app';
 import Vitessce from './Vitessce';
 import { encodeConfInUrl, decodeURLParamsToConf } from './export-utils';
-import { registerPluginFileType, registerPluginViewType } from './plugins';
+import { registerPluginCoordinationType, registerPluginViewType } from './plugins';
 
 export {
   createApp,
   Vitessce,
   encodeConfInUrl,
   decodeURLParamsToConf,
-  registerPluginFileType,
   registerPluginViewType,
+  registerPluginCoordinationType,
 };